requirements.txt
Flask==2.2.2
numpy==1.24.1
opencv_python==4.7.0.68
Pillow==9.4.0
pytesseract==0.3.10
